Refactor recording editor into windowed multi-segment flow
- Split the editor UI into window submodules and shared support
- Add cut/segment state with multi-segment FFmpeg export support
- Simplify onboarding and CSS to use the shared window chrome
